fix: refacto on search-drawer
