+++ src/components/custom/footer.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import SearchDrawer from "@/components/custom/search/search-drawer"
+import SearchWrapper from "@/components/custom/search/search-wrapper"
 import { Drawer, DrawerTrigger } from "@/components/ui/drawer"
 import Link from "next/link"
 import { useState } from "react"
@@ -20,7 +20,7 @@ export default function Footer() {
           <DrawerTrigger>
             <FiSearch className="w-6 h-6 text-gray-400 hover:text-green-500" />
           </DrawerTrigger>
-          <SearchDrawer setIsOpen={setIsOpen} />
+          <SearchWrapper setIsOpen={setIsOpen} />
         </Drawer>
         <button>
           <PiFlower className="w-6 h-6 text-gray-400 hover:text-green-500" />
